build: [#42309] Runner with Kubernetes executor
This change removes the explicit tags that forces these jobs to run on
the deprecated Gitlab Runners with the Docker executor. Without this, the
jobs will run on runners with the "Run untagged jobs" configuration.
Currently the only runner with this setting is the Gitlab Runner with
the Kubernetes executor.